remove studio related properties during command creation to prevent error when createCommandFromObject called outside of compiler
